fix: address reviews - rename `proof_expr` dir to `proof_exprs` - rename `proof_plan` dir to `proof_plans` - misc changes related to reviews
