+++ crates/proof-of-sql/src/sql/proof_exprs/add_subtract_expr_test.rs
@@ -7,8 +7,8 @@ use crate::{
     sql::{
         parse::ConversionError,
         proof::{exercise_verification, QueryError, VerifiableQueryResult},
-        proof_expr::{test_utility::*, DynProofExpr, ProofExpr},
-        proof_plan::{test_utility::*, DynProofPlan},
+        proof_exprs::{test_utility::*, DynProofExpr, ProofExpr},
+        proof_plans::{test_utility::*, DynProofPlan},
     },
 };
 use bumpalo::Bump;
@@ -20,50 +20,81 @@ use rand::{
 };
 use rand_core::SeedableRng;
 
-// select a * 2 as a, c, b * 4.5 as b, d * 3  + 4.7 as d, e from sxt.t where d * 3.9 = 8.19
+// select a, c, b + 4 as res, d from sxt.t where a - b = 3
 #[test]
-fn we_can_prove_a_typical_multiply_query() {
+fn we_can_prove_a_typical_add_subtract_query() {
     let data = owned_table([
         smallint("a", [1_i16, 2, 3, 4]),
-        int("b", [0_i32, 1, 2, 1]),
-        varchar("e", ["ab", "t", "efg", "g"]),
+        int("b", [0_i32, 1, 0, 1]),
+        varchar("d", ["ab", "t", "efg", "g"]),
         bigint("c", [0_i64, 2, 2, 0]),
-        decimal75("d", 2, 1, [21_i64, 4, 21, -7]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast = dense_filter(
         vec![
-            aliased_plan(multiply(column(t, "a", &accessor), const_int(2)), "a"),
+            col_expr_plan(t, "a", &accessor),
             col_expr_plan(t, "c", &accessor),
-            aliased_plan(
-                multiply(column(t, "b", &accessor), const_decimal75(2, 1, 45)),
-                "b",
-            ),
-            aliased_plan(
-                add(
-                    multiply(column(t, "d", &accessor), const_smallint(3)),
-                    const_decimal75(2, 1, 47),
-                ),
-                "d",
-            ),
-            col_expr_plan(t, "e", &accessor),
+            aliased_plan(add(column(t, "b", &accessor), const_bigint(4)), "res"),
+            col_expr_plan(t, "d", &accessor),
         ],
         tab(t),
         equal(
-            multiply(column(t, "d", &accessor), const_decimal75(2, 1, 39)),
-            const_decimal75(3, 2, 819),
+            subtract(column(t, "a", &accessor), column(t, "b", &accessor)),
+            const_bigint(3),
         ),
     );
     let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
     let expected_res = owned_table([
-        int("a", [2_i32, 6]),
-        bigint("c", [0_i64, 2]),
-        decimal75("b", 13, 1, [0_i64, 90]),
-        decimal75("d", 9, 1, [110_i64, 110]),
-        varchar("e", ["ab", "efg"]),
+        smallint("a", [3_i16, 4]),
+        bigint("c", [2_i16, 0]),
+        bigint("res", [4_i64, 5]),
+        varchar("d", ["efg", "g"]),
+    ]);
+    assert_eq!(res, expected_res);
+}
+
+// select a, a + b + c + 0.4 as c, d from sxt.t where a - b = 0.5
+#[test]
+fn we_can_prove_a_typical_add_subtract_query_with_decimals() {
+    let data = owned_table([
+        decimal75("a", 12, 1, [4_i64, 2, 2, 7]),
+        decimal75("b", 12, 2, [5_i64, -15, 42, 8]),
+        varchar("d", ["ab", "t", "efg", "g"]),
+        decimal75("c", 12, 3, [190_i64, 27, 253, 120]),
+    ]);
+    let t = "sxt.t".parse().unwrap();
+    let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
+    let ast = dense_filter(
+        vec![
+            col_expr_plan(t, "a", &accessor),
+            aliased_plan(
+                add(
+                    add(
+                        add(column(t, "a", &accessor), column(t, "b", &accessor)),
+                        column(t, "c", &accessor),
+                    ),
+                    const_decimal75(2, 1, 4),
+                ),
+                "c",
+            ),
+            col_expr_plan(t, "d", &accessor),
+        ],
+        tab(t),
+        equal(
+            subtract(column(t, "a", &accessor), column(t, "b", &accessor)),
+            const_decimal75(12, 4, 3500),
+        ),
+    );
+    let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
+    exercise_verification(&verifiable_res, &ast, &accessor, t);
+    let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
+    let expected_res = owned_table([
+        decimal75("a", 12, 1, [4_i64, 2]),
+        decimal75("c", 17, 3, [1040_i64, 477]),
+        varchar("d", ["ab", "t"]),
     ]);
     assert_eq!(res, expected_res);
 }
@@ -71,35 +102,32 @@ fn we_can_prove_a_typical_multiply_query() {
 // Column type issue tests
 #[test]
 fn decimal_column_type_issues_error_out_when_producing_provable_ast() {
-    let data = owned_table([decimal75("a", 57, 2, [1_i16, 2, 3, 4])]);
+    let data = owned_table([decimal75("a", 75, 2, [1_i16, 2, 3, 4])]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     assert!(matches!(
-        DynProofExpr::try_new_multiply(
-            column(t, "a", &accessor),
-            const_bigint::<RistrettoPoint>(1)
-        ),
+        DynProofExpr::try_new_add(column(t, "a", &accessor), const_bigint::<RistrettoPoint>(1)),
         Err(ConversionError::DataTypeMismatch(..))
     ));
 }
 
 // Overflow tests
-// select a * b as c from sxt.t where b = 2
+// select a + b as c from sxt.t where b = 1
 #[test]
 fn result_expr_can_overflow() {
     let data = owned_table([
         smallint("a", [i16::MAX, i16::MIN]),
-        smallint("b", [2_i16, 0]),
+        smallint("b", [1_i16, 0]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast: DynProofPlan<RistrettoPoint> = dense_filter(
         vec![aliased_plan(
-            multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
+            add(column(t, "a", &accessor), column(t, "b", &accessor)),
             "c",
         )],
         tab(t),
-        equal(column(t, "b", &accessor), const_bigint(2)),
+        equal(column(t, "b", &accessor), const_bigint(1)),
     );
     let verifiable_res: VerifiableQueryResult<InnerProductProof> =
         VerifiableQueryResult::new(&ast, &accessor, &());
@@ -109,18 +137,18 @@ fn result_expr_can_overflow() {
     ));
 }
 
-// select a * b as c from sxt.t where b == 0
+// select a + b as c from sxt.t where b == 0
 #[test]
 fn overflow_in_nonselected_rows_doesnt_error_out() {
     let data = owned_table([
         smallint("a", [i16::MAX, i16::MIN + 1]),
-        smallint("b", [2_i16, 0]),
+        smallint("b", [1_i16, 0]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast: DynProofPlan<RistrettoPoint> = dense_filter(
         vec![aliased_plan(
-            multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
+            add(column(t, "a", &accessor), column(t, "b", &accessor)),
             "c",
         )],
         tab(t),
@@ -130,24 +158,21 @@ fn overflow_in_nonselected_rows_doesnt_error_out() {
         VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([smallint("c", [0_i16])]);
+    let expected_res = owned_table([smallint("c", [i16::MIN + 1])]);
     assert_eq!(res, expected_res);
 }
 
-// select a, b from sxt.t where a * b >= 0
+// select a, b from sxt.t where a + b >= 0
 #[test]
 fn overflow_in_where_clause_doesnt_error_out() {
-    let data = owned_table([
-        bigint("a", [i64::MAX, i64::MIN + 1]),
-        smallint("b", [2_i16, 1]),
-    ]);
+    let data = owned_table([bigint("a", [i64::MAX, i64::MIN]), smallint("b", [1_i16, 0])]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast: DynProofPlan<RistrettoPoint> = dense_filter(
         cols_expr_plan(t, &["a", "b"], &accessor),
         tab(t),
         gte(
-            multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
+            add(column(t, "a", &accessor), column(t, "b", &accessor)),
             const_bigint(0),
         ),
     );
@@ -155,11 +180,11 @@ fn overflow_in_where_clause_doesnt_error_out() {
         VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([bigint("a", [i64::MAX]), smallint("b", [2_i16])]);
+    let expected_res = owned_table([bigint("a", [i64::MAX]), smallint("b", [1_i16])]);
     assert_eq!(res, expected_res);
 }
 
-// select a * b as c from sxt.t
+// select a + b as c, a - b as d from sxt.t
 #[test]
 fn result_expr_can_overflow_more() {
     let data = owned_table([
@@ -169,10 +194,16 @@ fn result_expr_can_overflow_more() {
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast: DynProofPlan<RistrettoPoint> = dense_filter(
-        vec![aliased_plan(
-            multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
-            "c",
-        )],
+        vec![
+            aliased_plan(
+                add(column(t, "a", &accessor), column(t, "b", &accessor)),
+                "c",
+            ),
+            aliased_plan(
+                subtract(column(t, "a", &accessor), column(t, "b", &accessor)),
+                "d",
+            ),
+        ],
         tab(t),
         const_bool(true),
     );
@@ -182,54 +213,6 @@ fn result_expr_can_overflow_more() {
         verifiable_res.verify(&ast, &accessor, &()),
         Err(QueryError::Overflow)
     ));
-}
-
-// select * from sxt.t where a * b * c * d * e = res
-// Only the last row is a valid result
-// The other two are due to the fact that scalars are elements of finite fields
-// and that hence scalar multiplication inherently wraps around
-#[test]
-fn where_clause_can_wrap_around() {
-    let data = owned_table([
-        bigint("a", [2357878470324616199_i64, 2657439699204141, 884]),
-        bigint("b", [31194601778911687_i64, 1644425323726039, 884]),
-        bigint("c", [500213946116239_i64, 1570568673569987, 884]),
-        bigint("d", [211980999383887_i64, 1056107792886999, 884]),
-        bigint("e", [927908842441_i64, 998426626609497, 884]),
-        bigint("res", [-20_i64, 50, 539835356263424]),
-    ]);
-    let t = "sxt.t".parse().unwrap();
-    let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
-    let ast: DynProofPlan<RistrettoPoint> = dense_filter(
-        cols_expr_plan(t, &["a", "b", "c", "d", "e", "res"], &accessor),
-        tab(t),
-        equal(
-            multiply(
-                multiply(
-                    multiply(
-                        multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
-                        column(t, "c", &accessor),
-                    ),
-                    column(t, "d", &accessor),
-                ),
-                column(t, "e", &accessor),
-            ),
-            column(t, "res", &accessor),
-        ),
-    );
-    let verifiable_res: VerifiableQueryResult<InnerProductProof> =
-        VerifiableQueryResult::new(&ast, &accessor, &());
-    exercise_verification(&verifiable_res, &ast, &accessor, t);
-    let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([
-        bigint("a", [2357878470324616199_i64, 2657439699204141, 884]),
-        bigint("b", [31194601778911687_i64, 1644425323726039, 884]),
-        bigint("c", [500213946116239_i64, 1570568673569987, 884]),
-        bigint("d", [211980999383887_i64, 1056107792886999, 884]),
-        bigint("e", [927908842441_i64, 998426626609497, 884]),
-        bigint("res", [-20_i64, 50, 539835356263424]),
-    ]);
-    assert_eq!(res, expected_res);
 }
 
 fn test_random_tables_with_given_offset(offset: usize) {
@@ -267,8 +250,8 @@ fn test_random_tables_with_given_offset(offset: usize) {
             vec![
                 col_expr_plan(t, "d", &accessor),
                 aliased_plan(
-                    add(
-                        multiply(column(t, "a", &accessor), column(t, "c", &accessor)),
+                    subtract(
+                        add(column(t, "a", &accessor), column(t, "c", &accessor)),
                         const_int128(4),
                     ),
                     "f",
@@ -296,7 +279,7 @@ fn test_random_tables_with_given_offset(offset: usize) {
         ))
         .filter_map(|(a, b, c, d)| {
             if b == &filter_val1 && c == &filter_val2 {
-                Some(((*a * *c + 4) as i128, d.clone()))
+                Some(((*a + *c - 4) as i128, d.clone()))
             } else {
                 None
             }
@@ -315,27 +298,27 @@ fn we_can_query_random_tables_using_a_zero_offset() {
 
 #[test]
 fn we_can_query_random_tables_using_a_non_zero_offset() {
-    test_random_tables_with_given_offset(23);
+    test_random_tables_with_given_offset(123);
 }
 
-// b * (a - 1.5)
+// b + a - 1
 #[test]
-fn we_can_compute_the_correct_output_of_a_multiply_expr_using_result_evaluate() {
+fn we_can_compute_the_correct_output_of_an_add_subtract_expr_using_result_evaluate() {
     let data = owned_table([
         smallint("a", [1_i16, 2, 3, 4]),
-        int("b", [0_i32, 1, 5, 1]),
+        int("b", [0_i32, 1, 0, 1]),
         varchar("d", ["ab", "t", "efg", "g"]),
         bigint("c", [0_i64, 2, 2, 0]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
-    let arithmetic_expr: DynProofExpr<RistrettoPoint> = multiply(
+    let add_subtract_expr: DynProofExpr<RistrettoPoint> = add(
         column(t, "b", &accessor),
-        subtract(column(t, "a", &accessor), const_decimal75(2, 1, 15)),
+        subtract(column(t, "a", &accessor), const_bigint(1)),
     );
     let alloc = Bump::new();
-    let res = arithmetic_expr.result_evaluate(4, &alloc, &accessor);
-    let expected_res_scalar = [0, 5, 75, 25]
+    let res = add_subtract_expr.result_evaluate(4, &alloc, &accessor);
+    let expected_res_scalar = [0, 2, 2, 4]
         .iter()
         .map(|v| Curve25519Scalar::from(*v))
         .collect::<Vec<_>>();
