+++ crates/proof-of-sql/src/sql/proof_exprs/aggregate_expr.rs
