+++ crates/proof-of-sql/src/sql/proof_exprs/aliased_dyn_proof_expr.rs
