+++ crates/proof-of-sql/src/sql/proof_exprs/and_expr.rs
