+++ crates/proof-of-sql/src/sql/proof_exprs/and_expr_test.rs
@@ -1,12 +1,12 @@
 use crate::{
     base::{
         commitment::InnerProductProof,
-        database::{owned_table_utility::*, Column, OwnedTableTestAccessor, TestAccessor},
+        database::{owned_table_utility::*, Column, OwnedTableTestAccessor},
     },
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
-        proof_expr::{test_utility::*, DynProofExpr, ProofExpr},
-        proof_plan::test_utility::*,
+        proof_exprs::{test_utility::*, DynProofExpr, ProofExpr},
+        proof_plans::test_utility::*,
     },
 };
 use bumpalo::Bump;
@@ -19,75 +19,52 @@ use rand::{
 use rand_core::SeedableRng;
 
 #[test]
-fn we_can_prove_a_simple_or_query() {
+fn we_can_prove_a_simple_and_query() {
     let data = owned_table([
-        bigint("a", [1_i64, 2, 3, 4]),
-        varchar("d", ["ab", "t", "g", "efg"]),
-        bigint("b", [0_i64, 1, 0, 2]),
+        bigint("a", [1, 2, 3, 4]),
+        bigint("b", [0, 1, 0, 1]),
+        varchar("d", ["ab", "t", "efg", "g"]),
+        bigint("c", [0, 2, 2, 0]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast = dense_filter(
         cols_expr_plan(t, &["a", "d"], &accessor),
         tab(t),
-        or(
-            equal(column(t, "b", &accessor), const_bigint(1)),
-            equal(column(t, "d", &accessor), const_varchar("g")),
+        and(
+            equal(column(t, "b", &accessor), const_scalar(1)),
+            equal(column(t, "d", &accessor), const_scalar("t")),
         ),
     );
     let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([bigint("a", [2_i64, 3]), varchar("d", ["t", "g"])]);
+    let expected_res = owned_table([bigint("a", [2]), varchar("d", ["t"])]);
     assert_eq!(res, expected_res);
 }
 
 #[test]
-fn we_can_prove_a_simple_or_query_with_variable_integer_types() {
+fn we_can_prove_a_simple_and_query_with_128_bits() {
     let data = owned_table([
-        int128("a", [1_i128, 2, 3, 4]),
-        varchar("d", ["ab", "t", "g", "efg"]),
-        smallint("b", [0_i16, 1, 0, 2]),
+        int128("a", [1, 2, 3, 4]),
+        int128("b", [0, 1, 0, 1]),
+        varchar("d", ["ab", "t", "efg", "g"]),
+        int128("c", [0, 2, 2, 0]),
     ]);
     let t = "sxt.t".parse().unwrap();
     let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
     let ast = dense_filter(
         cols_expr_plan(t, &["a", "d"], &accessor),
         tab(t),
-        or(
-            equal(column(t, "b", &accessor), const_bigint(1)),
-            equal(column(t, "d", &accessor), const_varchar("g")),
+        and(
+            equal(column(t, "b", &accessor), const_scalar(1)),
+            equal(column(t, "d", &accessor), const_scalar("t")),
         ),
     );
     let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([int128("a", [2_i64, 3]), varchar("d", ["t", "g"])]);
-    assert_eq!(res, expected_res);
-}
-
-#[test]
-fn we_can_prove_an_or_query_where_both_lhs_and_rhs_are_true() {
-    let data = owned_table([
-        bigint("a", [1_i64, 2, 3, 4]),
-        int128("b", [0_i128, 1, 1, 1]),
-        int("c", [0_i32, 2, 2, 0]),
-        varchar("d", ["ab", "t", "g", "efg"]),
-    ]);
-    let t = "sxt.t".parse().unwrap();
-    let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
-    let ast = dense_filter(
-        cols_expr_plan(t, &["a", "d"], &accessor),
-        tab(t),
-        or(
-            equal(column(t, "b", &accessor), const_bigint(1)),
-            equal(column(t, "d", &accessor), const_varchar("g")),
-        ),
-    );
-    let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
-    exercise_verification(&verifiable_res, &ast, &accessor, t);
-    let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected_res = owned_table([bigint("a", [2_i64, 3, 4]), varchar("d", ["t", "g", "efg"])]);
+    let expected_res = owned_table([int128("a", [2]), varchar("d", ["t"])]);
     assert_eq!(res, expected_res);
 }
 
@@ -125,7 +102,7 @@ fn test_random_tables_with_given_offset(offset: usize) {
         let ast = dense_filter(
             cols_expr_plan(t, &["a", "d"], &accessor),
             tab(t),
-            or(
+            and(
                 equal(
                     column(t, "b", &accessor),
                     const_varchar(filter_val1.as_str()),
@@ -145,7 +122,7 @@ fn test_random_tables_with_given_offset(offset: usize) {
             data["d"].string_iter(),
         ))
         .filter_map(|(a, b, c, d)| {
-            if b == &filter_val1 || c == &filter_val2 {
+            if b == &filter_val1 && c == &filter_val2 {
                 Some((*a, d.clone()))
             } else {
                 None
@@ -159,32 +136,31 @@ fn test_random_tables_with_given_offset(offset: usize) {
 }
 
 #[test]
-fn we_can_query_random_tables_with_a_zero_offset() {
+fn we_can_query_random_tables_using_a_zero_offset() {
     test_random_tables_with_given_offset(0);
 }
 
 #[test]
-fn we_can_query_random_tables_with_a_non_zero_offset() {
-    test_random_tables_with_given_offset(1001);
+fn we_can_query_random_tables_using_a_non_zero_offset() {
+    test_random_tables_with_given_offset(123);
 }
 
 #[test]
-fn we_can_compute_the_correct_output_of_an_or_expr_using_result_evaluate() {
+fn we_can_compute_the_correct_output_of_an_and_expr_using_result_evaluate() {
     let data = owned_table([
         bigint("a", [1, 2, 3, 4]),
         bigint("b", [0, 1, 0, 1]),
+        varchar("d", ["ab", "t", "efg", "g"]),
         bigint("c", [0, 2, 2, 0]),
-        varchar("d", ["ab", "t", "g", "efg"]),
     ]);
-    let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     let t = "sxt.t".parse().unwrap();
-    accessor.add_table(t, data, 0);
-    let and_expr: DynProofExpr<RistrettoPoint> = or(
+    let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
+    let and_expr: DynProofExpr<RistrettoPoint> = and(
         equal(column(t, "b", &accessor), const_int128(1)),
-        equal(column(t, "d", &accessor), const_varchar("g")),
+        equal(column(t, "d", &accessor), const_varchar("t")),
     );
     let alloc = Bump::new();
     let res = and_expr.result_evaluate(4, &alloc, &accessor);
-    let expected_res = Column::Boolean(&[false, true, true, true]);
+    let expected_res = Column::Boolean(&[false, true, false, false]);
     assert_eq!(res, expected_res);
 }
