+++ crates/proof-of-sql/src/sql/proof_exprs/bitwise_verification_test.rs
