+++ crates/proof-of-sql/src/sql/proof_exprs/column_expr.rs
