+++ crates/proof-of-sql/src/sql/proof_exprs/comparison_util.rs
