+++ crates/proof-of-sql/src/sql/proof_exprs/equals_expr.rs
