+++ crates/proof-of-sql/src/sql/proof_exprs/inequality_expr.rs
