+++ crates/proof-of-sql/src/sql/proof_exprs/literal_expr_test.rs
@@ -6,8 +6,8 @@ use crate::{
     },
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
-        proof_expr::test_utility::*,
-        proof_plan::test_utility::*,
+        proof_exprs::test_utility::*,
+        proof_plans::test_utility::*,
     },
 };
 use bumpalo::Bump;
