+++ crates/proof-of-sql/src/sql/proof_exprs/not_expr.rs
