+++ crates/proof-of-sql/src/sql/proof_exprs/numerical_util.rs
