+++ crates/proof-of-sql/src/sql/proof_exprs/sign_expr.rs
