+++ crates/proof-of-sql/src/sql/proof_exprs/table_expr.rs
