+++ crates/proof-of-sql/src/sql/proof_plans/dense_filter_exec_test.rs
@@ -1,8 +1,8 @@
-use super::{test_utility::*, DynProofPlan, ProjectionExec};
+use super::{test_utility::*, DenseFilterExec};
 use crate::{
     base::{
         database::{
-            owned_table_utility::*, ColumnField, ColumnRef, ColumnType, OwnedTable,
+            owned_table_utility::*, ColumnField, ColumnRef, ColumnType, LiteralValue, OwnedTable,
             OwnedTableTestAccessor, TableRef, TestAccessor,
         },
         math::decimal::Precision,
@@ -12,7 +12,7 @@ use crate::{
         proof::{
             exercise_verification, ProofPlan, ProverEvaluate, ResultBuilder, VerifiableQueryResult,
         },
-        proof_expr::{test_utility::*, ColumnExpr, DynProofExpr, TableExpr},
+        proof_exprs::{test_utility::*, ColumnExpr, DynProofExpr, LiteralExpr, TableExpr},
     },
 };
 use blitzar::proof::InnerProductProof;
@@ -26,7 +26,7 @@ fn we_can_correctly_fetch_the_query_result_schema() {
     let table_ref = TableRef::new(ResourceId::try_new("sxt", "sxt_tab").unwrap());
     let a = Identifier::try_new("a").unwrap();
     let b = Identifier::try_new("b").unwrap();
-    let provable_ast = ProjectionExec::<RistrettoPoint>::new(
+    let provable_ast = DenseFilterExec::<RistrettoPoint>::new(
         vec![
             aliased_plan(
                 DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
@@ -46,13 +46,23 @@ fn we_can_correctly_fetch_the_query_result_schema() {
             ),
         ],
         TableExpr { table_ref },
+        DynProofExpr::try_new_equals(
+            DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
+                table_ref,
+                Identifier::try_new("c").unwrap(),
+                ColumnType::BigInt,
+            ))),
+            DynProofExpr::Literal(LiteralExpr::new(LiteralValue::BigInt(123))),
+        )
+        .unwrap(),
     );
+
     let column_fields: Vec<ColumnField> = provable_ast.get_column_result_fields();
     assert_eq!(
         column_fields,
         vec![
             ColumnField::new("a".parse().unwrap(), ColumnType::BigInt),
-            ColumnField::new("b".parse().unwrap(), ColumnType::BigInt),
+            ColumnField::new("b".parse().unwrap(), ColumnType::BigInt)
         ]
     );
 }
@@ -62,7 +72,7 @@ fn we_can_correctly_fetch_all_the_referenced_columns() {
     let table_ref = TableRef::new(ResourceId::try_new("sxt", "sxt_tab").unwrap());
     let a = Identifier::try_new("a").unwrap();
     let f = Identifier::try_new("f").unwrap();
-    let provable_ast = ProjectionExec::<RistrettoPoint>::new(
+    let provable_ast = DenseFilterExec::new(
         vec![
             aliased_plan(
                 DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
@@ -82,6 +92,37 @@ fn we_can_correctly_fetch_all_the_referenced_columns() {
             ),
         ],
         TableExpr { table_ref },
+        not::<RistrettoPoint>(and(
+            or(
+                DynProofExpr::try_new_equals(
+                    DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
+                        table_ref,
+                        Identifier::try_new("f").unwrap(),
+                        ColumnType::BigInt,
+                    ))),
+                    DynProofExpr::Literal(LiteralExpr::new(LiteralValue::BigInt(45))),
+                )
+                .unwrap(),
+                DynProofExpr::try_new_equals(
+                    DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
+                        table_ref,
+                        Identifier::try_new("c").unwrap(),
+                        ColumnType::BigInt,
+                    ))),
+                    DynProofExpr::Literal(LiteralExpr::new(LiteralValue::BigInt(-2))),
+                )
+                .unwrap(),
+            ),
+            DynProofExpr::try_new_equals(
+                DynProofExpr::Column(ColumnExpr::new(ColumnRef::new(
+                    table_ref,
+                    Identifier::try_new("b").unwrap(),
+                    ColumnType::BigInt,
+                ))),
+                DynProofExpr::Literal(LiteralExpr::new(LiteralValue::BigInt(3))),
+            )
+            .unwrap(),
+        )),
     );
 
     let ref_columns = provable_ast.get_column_references();
@@ -99,58 +140,39 @@ fn we_can_correctly_fetch_all_the_referenced_columns() {
                 Identifier::try_new("f").unwrap(),
                 ColumnType::BigInt
             ),
+            ColumnRef::new(
+                table_ref,
+                Identifier::try_new("c").unwrap(),
+                ColumnType::BigInt
+            ),
+            ColumnRef::new(
+                table_ref,
+                Identifier::try_new("b").unwrap(),
+                ColumnType::BigInt
+            )
         ])
     );
 }
 
 #[test]
-fn we_can_prove_and_get_the_correct_result_from_a_basic_projection() {
-    let data = owned_table([
-        bigint("a", [1_i64, 4_i64, 5_i64, 2_i64, 5_i64, 1, 4, 5, 2, 5]),
-        bigint("b", [1_i64, 2, 3, 4, 5, 1, 2, 3, 4, 5]),
-    ]);
-    let t = "sxt.t".parse().unwrap();
-    let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
-    accessor.add_table(t, data, 0);
-    let ast = projection(cols_expr_plan(t, &["b"], &accessor), tab(t));
-    let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
-    exercise_verification(&verifiable_res, &ast, &accessor, t);
-    let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected = owned_table([bigint("b", [1_i64, 2, 3, 4, 5, 1, 2, 3, 4, 5])]);
-    assert_eq!(res, expected);
-}
-
-#[test]
-fn we_can_prove_and_get_the_correct_result_from_a_nontrivial_projection() {
+fn we_can_prove_and_get_the_correct_result_from_a_basic_dense_filter() {
     let data = owned_table([
         bigint("a", [1_i64, 4_i64, 5_i64, 2_i64, 5_i64]),
         bigint("b", [1_i64, 2, 3, 4, 5]),
     ]);
     let t = "sxt.t".parse().unwrap();
-    let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
-    accessor.add_table(t, data, 0);
-    let ast = projection(
-        vec![
-            aliased_plan(add(column(t, "b", &accessor), const_bigint(1)), "b"),
-            aliased_plan(
-                multiply(column(t, "a", &accessor), column(t, "b", &accessor)),
-                "prod",
-            ),
-        ],
-        tab(t),
-    );
+    let accessor = OwnedTableTestAccessor::<InnerProductProof>::new_from_table(t, data, 0, ());
+    let where_clause = equal(column(t, "a", &accessor), const_int128(5_i128));
+    let ast = dense_filter(cols_expr_plan(t, &["b"], &accessor), tab(t), where_clause);
     let verifiable_res = VerifiableQueryResult::new(&ast, &accessor, &());
     exercise_verification(&verifiable_res, &ast, &accessor, t);
     let res = verifiable_res.verify(&ast, &accessor, &()).unwrap().table;
-    let expected = owned_table([
-        bigint("b", [2_i64, 3, 4, 5, 6]),
-        bigint("prod", [1_i64, 8, 15, 8, 25]),
-    ]);
-    assert_eq!(res, expected);
+    let expected_res = owned_table([bigint("b", [3_i64, 5])]);
+    assert_eq!(res, expected_res);
 }
 
 #[test]
-fn we_can_get_an_empty_result_from_a_basic_projection_on_an_empty_table_using_result_evaluate() {
+fn we_can_get_an_empty_result_from_a_basic_dense_filter_on_an_empty_table_using_result_evaluate() {
     let data = owned_table([
         bigint("a", [0; 0]),
         bigint("b", [0; 0]),
@@ -161,8 +183,13 @@ fn we_can_get_an_empty_result_from_a_basic_projection_on_an_empty_table_using_re
     let t = "sxt.t".parse().unwrap();
     let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     accessor.add_table(t, data, 0);
-    let expr: DynProofPlan<RistrettoPoint> =
-        projection(cols_expr_plan(t, &["b", "c", "d", "e"], &accessor), tab(t));
+    let where_clause: DynProofExpr<RistrettoPoint> =
+        equal(column(t, "a", &accessor), const_int128(999));
+    let expr = dense_filter(
+        cols_expr_plan(t, &["b", "c", "d", "e"], &accessor),
+        tab(t),
+        where_clause,
+    );
     let alloc = Bump::new();
     let mut builder = ResultBuilder::new(0);
     expr.result_evaluate(&mut builder, &alloc, &accessor);
@@ -190,7 +217,7 @@ fn we_can_get_an_empty_result_from_a_basic_projection_on_an_empty_table_using_re
 }
 
 #[test]
-fn we_can_get_no_columns_from_a_basic_projection_with_no_selected_columns_using_result_evaluate() {
+fn we_can_get_an_empty_result_from_a_basic_dense_filter_using_result_evaluate() {
     let data = owned_table([
         bigint("a", [1, 4, 5, 2, 5]),
         bigint("b", [1, 2, 3, 4, 5]),
@@ -201,7 +228,55 @@ fn we_can_get_no_columns_from_a_basic_projection_with_no_selected_columns_using_
     let t = "sxt.t".parse().unwrap();
     let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     accessor.add_table(t, data, 0);
-    let expr: DynProofPlan<RistrettoPoint> = projection(cols_expr_plan(t, &[], &accessor), tab(t));
+    let where_clause: DynProofExpr<RistrettoPoint> =
+        equal(column(t, "a", &accessor), const_int128(999));
+    let expr = dense_filter(
+        cols_expr_plan(t, &["b", "c", "d", "e"], &accessor),
+        tab(t),
+        where_clause,
+    );
+    let alloc = Bump::new();
+    let mut builder = ResultBuilder::new(5);
+    expr.result_evaluate(&mut builder, &alloc, &accessor);
+    let fields = &[
+        ColumnField::new("b".parse().unwrap(), ColumnType::BigInt),
+        ColumnField::new("c".parse().unwrap(), ColumnType::Int128),
+        ColumnField::new("d".parse().unwrap(), ColumnType::VarChar),
+        ColumnField::new(
+            "e".parse().unwrap(),
+            ColumnType::Decimal75(Precision::new(1).unwrap(), 0),
+        ),
+    ];
+    let res = builder
+        .make_provable_query_result()
+        .to_owned_table(fields)
+        .unwrap();
+    let expected: OwnedTable<Curve25519Scalar> = owned_table([
+        bigint("b", [0; 0]),
+        int128("c", [0; 0]),
+        varchar("d", [""; 0]),
+        decimal75("e", 1, 0, [0; 0]),
+    ]);
+
+    assert_eq!(res, expected);
+}
+
+#[test]
+fn we_can_get_no_columns_from_a_basic_dense_filter_with_no_selected_columns_using_result_evaluate()
+{
+    let data = owned_table([
+        bigint("a", [1, 4, 5, 2, 5]),
+        bigint("b", [1, 2, 3, 4, 5]),
+        int128("c", [1, 2, 3, 4, 5]),
+        varchar("d", ["1", "2", "3", "4", "5"]),
+        scalar("e", [1, 2, 3, 4, 5]),
+    ]);
+    let t = "sxt.t".parse().unwrap();
+    let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
+    accessor.add_table(t, data, 0);
+    let where_clause: DynProofExpr<RistrettoPoint> =
+        equal(column(t, "a", &accessor), const_int128(5));
+    let expr = dense_filter(cols_expr_plan(t, &[], &accessor), tab(t), where_clause);
     let alloc = Bump::new();
     let mut builder = ResultBuilder::new(5);
     expr.result_evaluate(&mut builder, &alloc, &accessor);
@@ -215,7 +290,7 @@ fn we_can_get_no_columns_from_a_basic_projection_with_no_selected_columns_using_
 }
 
 #[test]
-fn we_can_get_the_correct_result_from_a_basic_projection_using_result_evaluate() {
+fn we_can_get_the_correct_result_from_a_basic_dense_filter_using_result_evaluate() {
     let data = owned_table([
         bigint("a", [1, 4, 5, 2, 5]),
         bigint("b", [1, 2, 3, 4, 5]),
@@ -226,24 +301,19 @@ fn we_can_get_the_correct_result_from_a_basic_projection_using_result_evaluate()
     let t = "sxt.t".parse().unwrap();
     let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     accessor.add_table(t, data, 0);
-    let expr: DynProofPlan<RistrettoPoint> = projection(
-        vec![
-            aliased_plan(add(column(t, "b", &accessor), const_bigint(1)), "b"),
-            aliased_plan(
-                multiply(column(t, "b", &accessor), column(t, "c", &accessor)),
-                "prod",
-            ),
-            col_expr_plan(t, "d", &accessor),
-            aliased_plan(const_decimal75(1, 0, 3), "e"),
-        ],
+    let where_clause: DynProofExpr<RistrettoPoint> =
+        equal(column(t, "a", &accessor), const_int128(5));
+    let expr = dense_filter(
+        cols_expr_plan(t, &["b", "c", "d", "e"], &accessor),
         tab(t),
+        where_clause,
     );
     let alloc = Bump::new();
     let mut builder = ResultBuilder::new(5);
     expr.result_evaluate(&mut builder, &alloc, &accessor);
     let fields = &[
         ColumnField::new("b".parse().unwrap(), ColumnType::BigInt),
-        ColumnField::new("prod".parse().unwrap(), ColumnType::Int128),
+        ColumnField::new("c".parse().unwrap(), ColumnType::Int128),
         ColumnField::new("d".parse().unwrap(), ColumnType::VarChar),
         ColumnField::new(
             "e".parse().unwrap(),
@@ -255,16 +325,16 @@ fn we_can_get_the_correct_result_from_a_basic_projection_using_result_evaluate()
         .to_owned_table(fields)
         .unwrap();
     let expected: OwnedTable<Curve25519Scalar> = owned_table([
-        bigint("b", [2, 3, 4, 5, 6]),
-        int128("prod", [1, 4, 9, 16, 25]),
-        varchar("d", ["1", "2", "3", "4", "5"]),
-        decimal75("e", 1, 0, [3; 5]),
+        bigint("b", [3, 5]),
+        int128("c", [3, 5]),
+        varchar("d", ["3", "5"]),
+        decimal75("e", 1, 0, [3, 5]),
     ]);
     assert_eq!(res, expected);
 }
 
 #[test]
-fn we_can_prove_a_projection_on_an_empty_table() {
+fn we_can_prove_a_dense_filter_on_an_empty_table() {
     let data = owned_table([
         bigint("a", [101; 0]),
         bigint("b", [3; 0]),
@@ -275,32 +345,54 @@ fn we_can_prove_a_projection_on_an_empty_table() {
     let t = "sxt.t".parse().unwrap();
     let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     accessor.add_table(t, data, 0);
-    let expr = projection(
-        vec![
-            aliased_plan(add(column(t, "b", &accessor), const_bigint(1)), "b"),
-            aliased_plan(
-                multiply(column(t, "b", &accessor), column(t, "c", &accessor)),
-                "prod",
-            ),
-            col_expr_plan(t, "d", &accessor),
-            aliased_plan(const_decimal75(1, 0, 3), "e"),
-        ],
+    let expr = dense_filter(
+        cols_expr_plan(t, &["b", "c", "d", "e"], &accessor),
         tab(t),
+        equal(column(t, "a", &accessor), const_int128(106)),
     );
     let res = VerifiableQueryResult::new(&expr, &accessor, &());
     exercise_verification(&res, &expr, &accessor, t);
     let res = res.verify(&expr, &accessor, &()).unwrap().table;
     let expected = owned_table([
         bigint("b", [3; 0]),
-        int128("prod", [3; 0]),
+        int128("c", [3; 0]),
         varchar("d", ["3"; 0]),
-        decimal75("e", 1, 0, [3; 0]),
+        scalar("e", [3; 0]),
     ]);
     assert_eq!(res, expected);
 }
 
 #[test]
-fn we_can_prove_a_projection() {
+fn we_can_prove_a_dense_filter_with_empty_results() {
+    let data = owned_table([
+        bigint("a", [101, 104, 105, 102, 105]),
+        bigint("b", [1, 2, 3, 4, 5]),
+        int128("c", [1, 2, 3, 4, 5]),
+        varchar("d", ["1", "2", "3", "4", "5"]),
+        scalar("e", [1, 2, 3, 4, 5]),
+    ]);
+    let t = "sxt.t".parse().unwrap();
+    let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
+    accessor.add_table(t, data, 0);
+    let expr = dense_filter(
+        cols_expr_plan(t, &["b", "c", "d", "e"], &accessor),
+        tab(t),
+        equal(column(t, "a", &accessor), const_int128(106)),
+    );
+    let res = VerifiableQueryResult::new(&expr, &accessor, &());
+    exercise_verification(&res, &expr, &accessor, t);
+    let res = res.verify(&expr, &accessor, &()).unwrap().table;
+    let expected = owned_table([
+        bigint("b", [3; 0]),
+        int128("c", [3; 0]),
+        varchar("d", ["3"; 0]),
+        scalar("e", [3; 0]),
+    ]);
+    assert_eq!(res, expected);
+}
+
+#[test]
+fn we_can_prove_a_dense_filter() {
     let data = owned_table([
         bigint("a", [101, 104, 105, 102, 105]),
         bigint("b", [1, 2, 3, 4, 7]),
@@ -311,7 +403,7 @@ fn we_can_prove_a_projection() {
     let t = "sxt.t".parse().unwrap();
     let mut accessor = OwnedTableTestAccessor::<InnerProductProof>::new_empty_with_setup(());
     accessor.add_table(t, data, 0);
-    let expr = projection(
+    let expr = dense_filter(
         vec![
             col_expr_plan(t, "b", &accessor),
             col_expr_plan(t, "c", &accessor),
@@ -324,17 +416,18 @@ fn we_can_prove_a_projection() {
             ),
         ],
         tab(t),
+        equal(column(t, "a", &accessor), const_int128(105)),
     );
     let res = VerifiableQueryResult::new(&expr, &accessor, &());
     exercise_verification(&res, &expr, &accessor, t);
     let res = res.verify(&expr, &accessor, &()).unwrap().table;
     let expected = owned_table([
-        bigint("b", [1, 2, 3, 4, 7]),
-        int128("c", [1, 3, 3, 4, 5]),
-        varchar("d", ["1", "2", "3", "4", "5"]),
-        scalar("e", [1, 2, 3, 4, 5]),
-        int128("const", [105; 5]),
-        boolean("bool", [true, false, true, true, false]),
+        bigint("b", [3, 7]),
+        int128("c", [3, 5]),
+        varchar("d", ["3", "5"]),
+        scalar("e", [3, 5]),
+        int128("const", [105, 105]),
+        boolean("bool", [true, false]),
     ]);
     assert_eq!(res, expected);
 }
