+++ crates/proof-of-sql/src/sql/proof_plans/dyn_proof_plan.rs
