+++ crates/proof-of-sql/src/sql/proof_plans/filter_exec_test_dishonest_prover.rs
@@ -10,7 +10,7 @@ use crate::{
             Indexes, ProofBuilder, ProverEvaluate, ProverHonestyMarker, QueryError, ResultBuilder,
             VerifiableQueryResult,
         },
-        proof_expr::{test_utility::*, ProofExpr},
+        proof_exprs::{test_utility::*, ProofExpr},
     },
 };
 use blitzar::proof::InnerProductProof;
