+++ crates/proof-of-sql/src/sql/proof_plans/filter_result_expr.rs
