+++ crates/proof-of-sql/src/sql/proof_plans/fold_util.rs
