+++ crates/proof-of-sql/src/sql/proof_plans/group_by_exec_test.rs
@@ -7,7 +7,7 @@ use crate::{
     },
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
-        proof_expr::test_utility::*,
+        proof_exprs::test_utility::*,
     },
 };
